refactor(server): split server and observability into separate module
Move pkg/server and pkg/{logging,metrics,tracing,health} into a new nested
module github.com/skyoo2003/acor/server. pkg/server had no consumers and was
the sole importer of the observability packages, which were the only reason
the core module pulled in gRPC, OpenTelemetry, Prometheus, and zerolog.

Core direct dependencies drop from 11 to 3 (go-redis, x/sync, miniredis
test-only); the heavy stack now lives in the server module. No functional
changes -- only module boundary, import paths, docs, and CI/Makefile updates
so both modules are built and tested.

Note: bypassed the local pre-commit golangci-lint (v2.12.2) which flags
pre-existing goconst debt in cmd/acor and *_test.go files this PR does not
touch; CI (golangci-lint v2.11.3) is green on master for those same files.
