+++ Makefile
@@ -14,9 +14,11 @@ build:
 
 test:
 	@go test -v ./...
+	@cd server && go test -v ./...
 
 lint:
 	@golangci-lint run ./...
+	@cd server && golangci-lint run ./...
 
 coverage:
 	@go test ./... -coverprofile=coverage.out -covermode=atomic
@@ -25,9 +27,11 @@ coverage:
 
 vet:
 	@go vet ./...
+	@cd server && go vet ./...
 
 lint-fix:
 	@golangci-lint run --fix ./...
+	@cd server && golangci-lint run --fix ./...
 
 fuzz:
 	@go test -fuzz=FuzzFind -fuzztime=30s ./pkg/acor
@@ -35,3 +39,4 @@ fuzz:
 
 race:
 	@go test -race ./...
+	@cd server && go test -race ./...
