+++ docs/content/_index.md
@@ -130,7 +130,9 @@ matches, _ := ac.FindParallel(largeText, &acor.ParallelOptions{
 
 ## Observability
 
-ACOR provides built-in observability packages:
+ACOR provides built-in observability packages in a separate module
+(`go get github.com/skyoo2003/acor/server`) so the core library stays
+dependency-light:
 
 - **Metrics**: Prometheus metrics for HTTP, gRPC, and Redis
 - **Logging**: Structured JSON logging with zerolog
@@ -139,10 +141,10 @@ ACOR provides built-in observability packages:
 
 ```go
 import (
-    "github.com/skyoo2003/acor/pkg/metrics"
-    "github.com/skyoo2003/acor/pkg/logging"
-    "github.com/skyoo2003/acor/pkg/tracing"
-    "github.com/skyoo2003/acor/pkg/health"
+    "github.com/skyoo2003/acor/server/metrics"
+    "github.com/skyoo2003/acor/server/logging"
+    "github.com/skyoo2003/acor/server/tracing"
+    "github.com/skyoo2003/acor/server/health"
 )
 ```
